requirements.txt
Django==2.2.1
pytz==2019.1
sqlparse==0.3.0
uwsgi
django-phonenumber-field
phonenumbers